Add files via upload
This program produces an Online Examination system, where sign-in and new registration options are available for users and the user details are stored in a database. Once the user signs in, the user can take the test for three subjects, each having 3 MCQs. The results are finally displayed with a Results.jsp page where the correct and wrong status of answers chosen are displayed along with the total percentage finally obtained.